+++ feature/notice/src/main/java/team/retum/jobis/notice/navigation/NoticeDetailsNavigation.kt
@@ -1,9 +1,9 @@
-package team.retum.jobis.notification.navigation
+package team.retum.jobis.notice.navigation
 
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
-import team.retum.jobis.notification.ui.NoticeDetails
+import team.retum.jobis.notice.ui.NoticeDetails
 
 const val NAVIGATION_NOTICE_DETAILS = "noticeDetails/"
 const val NOTICE_ID = "{notice-id}"
